improving the neighbor-based alignment to properly penalize non-alignment and reward alignment; also fixed an issue with the 'number of swaps' parametrization

+++ src/C/common.hpp
@@ -45,7 +45,7 @@ typedef struct {
         vector<int> match2;
         double (*dfunc)(Role*,Role*);
 
-        int iters_fixed_T;
+        double iters_fixed_T;
         double t_initial;
         double mu_t;
         double t_min;
